UserManagementSystem/requirements.txt
fastapi
uvicorn
streamlit
requests
pillow
pydantic
